Move before action for regenerate_api_key
Raising for missing callback actions is a new default in Rails 7.1. Therefore
a callback that is defined by an extended concern will cause a failure if the
action is not defined. The `regenerate_api_key` only exists for external user
admin so the callback needs to be moved from the concern to the controller.