Fix test for Ruby version of sum method
In Rails 7.1 the `sum` method no longer overrides `Enumerable#sum`. One change
is that there is no implicit conversion of nil to integer. When a fee is
cleared with `Fee::BaseFee#clear` the amount is set to nil and so
claim.basic_fees.sum(&:amount) fails.

It might be appropriate to change `Fee::BaseFee#clear` so that the values
are set to zero instead of nil. However, this may have other side-effects.