Update error regex
In two places, failing methods are identified in error messages using
regexes. Ruby 3.4 has changed the format of error messages so that they
now start with a single quote instead of a backtick.

https://bugs.ruby-lang.org/issues/16495

As a result these regexes now do not identify the methods correctly
resulting in failures. This is fixed by updating the regexes, replacing
backticks with single quotes.